refactor(price): rename fluxAggregatorKit.js
