+++ packages/inter-protocol/src/price/README.md
@@ -1,6 +1,6 @@
 # Price oracles
 
-This directory contains the `fluxAggregator.js` contract which takes prices as
+This directory contains the `fluxAggregatorKit.js` contract which takes prices as
 input and outputs a best known price. There are multiple ways to get the price,
 including a PriceAuthority interface.
