+++ packages/inter-protocol/test/price/snapshots/test-fluxAggregatorKit.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/price/test-fluxAggregator.js`
+# Snapshot report for `test/price/test-fluxAggregatorKit.js`
 
-The actual snapshot is saved in `test-fluxAggregator.js.snap`.
+The actual snapshot is saved in `test-fluxAggregatorKit.js.snap`.
 
 Generated by [AVA](https://avajs.dev).
