+++ packages/zoe/src/contracts/priceAggregator.js
@@ -47,7 +47,7 @@ const priceDescriptionFromQuote = quote => quote.quoteAmount.value[0];
  *
  * This contract aggregates price values from a set of oracles and provides a
  * PriceAuthority for their median. This naive method is game-able and so this module
- * is a stub until we complete what is now in `fluxAggregator.js`.
+ * is a stub until we complete what is now in `fluxAggregatorKit.js`.
  *
  * @param {ZCF<{
  * timer: import('@agoric/time/src/types').TimerService,
